CRUD finish

+++ src/app/heroes/pages/add/add.component.ts
@@ -1,6 +1,10 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
+import { MatSnackBar } from '@angular/material/snack-bar';
 import { ActivatedRoute, Router } from '@angular/router';
 import { switchMap } from 'rxjs/operators';
+
+import { DialogComponent } from '../../components/dialog/dialog.component';
 
 import { Heroe, Publisher } from '../../interfaces/heroes.interface';
 import { HeroesService } from '../../services/heroes.service';
@@ -37,10 +41,12 @@ export class AddComponent implements OnInit {
     alt_img:  ''
   };
 
-  constructor( 
+  constructor(
       private heroesService: HeroesService,
       private activatedRoute: ActivatedRoute,
-      private router: Router
+      private router: Router,
+      private snacKBar: MatSnackBar,
+      public dialog: MatDialog
   ) { }
 
   ngOnInit(): void {
@@ -61,17 +67,36 @@ export class AddComponent implements OnInit {
     if ( this.heroe.id ) {
       // Update
       this.heroesService.updateHeroe( this.heroe )
-        .subscribe( heroe => console.log( 'Update', heroe ) );
+        .subscribe( heroe => this.showSnackBar('Heroe updated') );
     } else {
       // Add
       this.heroesService.addHeroe( this.heroe )
-        .subscribe( heroe => this.router.navigate(['/heroes/edit', heroe.id ]));
+        .subscribe( heroe => {
+          this.router.navigate(['/heroes/edit', heroe.id ]);
+          this.showSnackBar('Heroe created');
+        });
     }
   }
 
-  delete() {
-    this.heroesService.deleteHeroe( this.heroe.id! )
-      .subscribe( resp => this.router.navigate(['/heroes']));
+  delete(): void {
+    const dialog = this.dialog.open( DialogComponent, {
+      width: '300px',
+      data: this.heroe
+    } );
+
+    dialog.afterClosed()
+      .subscribe( ( result ) => {
+        if ( result ) {
+        this.heroesService.deleteHeroe( this.heroe.id! )
+          .subscribe( resp => this.router.navigate(['/heroes']));
+        }
+      });
+  }
+
+  showSnackBar( message: string ): void {
+    this.snacKBar.open( message, 'Ok!', {
+      duration: 2500
+    });
   }
 
 }
